ARM: can build with llvm, switch to EABI

+++ minix/servers/is/dmp_pm.c
@@ -94,7 +94,8 @@ void sigaction_dmp()
 		mp->mp_ignore.__bits[0], mp->mp_catch.__bits[0],
 		mp->mp_sigmask.__bits[0]);
 	printf("%08x  ", mp->mp_sigpending.__bits[0]);
-  	if (mp->mp_flags & ALARM_ON) printf("%8lu", mp->mp_timer.tmr_exp_time-uptime);
+  	if (mp->mp_flags & ALARM_ON) printf("%8lu",
+		(unsigned long) (mp->mp_timer.tmr_exp_time-uptime));
   	else printf("       -");
   	printf("\n");
   }
